refactor(ai): introduce turn request builder module
