+++ src/node/services/aiService.ts
@@ -12,6 +12,8 @@ import type { Result } from "@/common/types/result";
 import { Ok, Err } from "@/common/types/result";
 import type { WorkspaceMetadata } from "@/common/types/workspace";
 import type { SendMessageOptions, ProvidersConfigMap } from "@/common/orpc/types";
+import type { StreamMessageOptions } from "./turnRequestBuilder";
+export type { StreamMessageOptions } from "./turnRequestBuilder";
 
 import type { DebugLlmRequestSnapshot } from "@/common/types/debugLlmRequest";
 import {
@@ -21,7 +23,7 @@ import {
 import { EXPERIMENT_IDS, type ExperimentId } from "@/common/constants/experiments";
 
 import type { GoalRecordV1 } from "@/common/types/goal";
-import type { ModelMessage, MuxMessage, MuxMessageMetadata } from "@/common/types/message";
+import type { ModelMessage, MuxMessage } from "@/common/types/message";
 import { createMuxMessage } from "@/common/types/message";
 import type { Config } from "@/node/config";
 import {
@@ -76,15 +78,12 @@ import type { PolicyService } from "@/node/services/policyService";
 import type { ProviderService } from "@/node/services/providerService";
 import type { CodexOauthService } from "@/node/services/codexOauthService";
 import type { CoderOauthService } from "@/node/services/coderOauthService";
-import type { WorkspaceGoalService } from "@/node/services/workspaceGoalService";
 import type { BackgroundProcessManager } from "@/node/services/backgroundProcessManager";
-import type { FileState } from "@/node/services/agentSession";
 import { log } from "./log";
 import {
   addInterruptedSentinel,
   filterEmptyAssistantMessages,
 } from "@/browser/utils/messages/modelMessageTransform";
-import type { PostCompactionAttachment } from "@/common/types/attachment";
 
 import type { HistoryService } from "./historyService";
 import { delegatedToolCallManager } from "./delegatedToolCallManager";
@@ -146,11 +145,7 @@ import { isWorkspaceTrustedForSharedExecution } from "@/node/services/utils/work
 import { DEFAULT_GOAL_DEFAULTS, normalizeGoalDefaults } from "@/constants/goals";
 import { mergeGoalDefaults } from "@/common/utils/goals/resolveGoalSetIntent";
 import { MULTI_PROJECT_CONFIG_KEY } from "@/common/constants/multiProject";
-import {
-  THINKING_LEVEL_OFF,
-  type OpenAIReasoningMode,
-  type ThinkingLevel,
-} from "@/common/types/thinking";
+import { THINKING_LEVEL_OFF, type ThinkingLevel } from "@/common/types/thinking";
 import {
   enforceThinkingPolicy,
   isXaiGrokFastVariantSwap,
@@ -159,13 +154,11 @@ import {
   resolveMinimumThinkingLevel,
 } from "@/common/utils/thinking/policy";
 import type {
-  ActiveTurnThinkingOverride,
   RebuildFirstStepForThinkingLevel,
   RebuildProviderOptionsForThinkingLevel,
 } from "@/node/services/thinkingOverride";
 
-import type { ErrorEvent, StreamAbortEvent, StreamAbortReason } from "@/common/types/stream";
-import type { ToolPolicy } from "@/common/utils/tools/toolPolicy";
+import type { StreamAbortEvent, StreamAbortReason } from "@/common/types/stream";
 import {
   computeActiveToolNames,
   prepareToolSearch,
@@ -273,70 +266,6 @@ export function replaceOrAppendMessageById(
 // ---------------------------------------------------------------------------
 // streamMessage options
 // ---------------------------------------------------------------------------
-
-/** Options bag for {@link AIService.streamMessage}. */
-export interface StreamMessageOptions {
-  messages: MuxMessage[];
-  workspaceId: string;
-  modelString: string;
-  thinkingLevel?: ThinkingLevel;
-  /** OpenAI pro reasoning mode; delivered via provider options (inert for unsupported models). */
-  reasoningMode?: OpenAIReasoningMode;
-  toolPolicy?: ToolPolicy;
-  abortSignal?: AbortSignal;
-  /** Live workspace scratchpad snapshot from the renderer; when present it wins over disk. */
-  additionalSystemContext?: string;
-  additionalSystemInstructions?: string;
-  maxOutputTokens?: number;
-  muxProviderOptions?: MuxProviderOptions;
-  /** Internal-only flag for Copilot billing attribution; never sourced from IPC schemas. */
-  agentInitiated?: boolean;
-  agentId?: string;
-  /** See SendMessageOptionsSchema.strictAgentResolution: explicit-agent sends fail loudly instead of falling back to exec. */
-  strictAgentResolution?: SendMessageOptions["strictAgentResolution"];
-  /** ACP prompt correlation id used to match stream events to a specific request. */
-  acpPromptId?: string;
-  /** Invoked with each fatal pre-start error event this call emits before returning Err. */
-  onPreStartError?: (event: ErrorEvent) => void;
-  /** Tool names that should be delegated back to ACP clients for this request. */
-  delegatedToolNames?: string[];
-  recordFileState?: (filePath: string, state: FileState) => Promise<void>;
-  postCompactionAttachments?: PostCompactionAttachment[] | null;
-  /**
-   * Resolver for the session-segment memory context (memory experiment):
-   * index snapshot for the memory tool description + hot-memories block.
-   * AgentSession caches the result per model/session segment because hot-memory
-   * selection is token-budgeted with the active model tokenizer. A callback
-   * (not a pre-resolved value) because it must be computed after
-   * runtime.ensureReady(): project-scope listing on a
-   * stopped Docker/remote workspace would otherwise cache an empty/partial
-   * context for the whole segment.
-   */
-  resolveMemoryContext?: (
-    modelString: string,
-    options?: { includeHotMemories?: boolean }
-  ) => Promise<MemorySessionContext | undefined>;
-  experiments?: SendMessageOptions["experiments"];
-  allowAgentSetGoal?: boolean;
-  workspaceGoalService?: WorkspaceGoalService;
-  disableWorkspaceAgents?: boolean;
-  hasQueuedMessages?: (dispatchMode?: "tool-end" | "turn-end") => boolean;
-  muxMetadata?: MuxMessageMetadata;
-  openaiTruncationModeOverride?: "auto" | "disabled";
-  /**
-   * Model floor already resolved by AgentSession (config.json
-   * minThinkingLevelByModel → resolveMinimumThinkingLevel). Passed down so
-   * mid-turn overrides clamp against the same floor as the send-time level;
-   * internal callers may omit it (re-resolved from defaults).
-   */
-  minThinkingLevel?: ThinkingLevel;
-  /**
-   * Session-owned per-turn holder for mid-turn thinking-level overrides.
-   * When absent (compaction, sub-agent paths), the feature is inert for the
-   * stream. See src/node/services/thinkingOverride.ts.
-   */
-  activeTurnThinkingOverride?: ActiveTurnThinkingOverride;
-}
 
 /**
  * Recursively merge user-provided provider extras under Xum-built provider options.
